backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
transformers>=4.2
torch>=1.12.1
djangorestframework>=3.13
geoip2==2.9.0
geopy==2.2.0
django-db-cascade==0.1.6
requests~=2.28.1